disk: remove local restriction
Currently, the disk script fails to scan a disk if it's over the network. If a
user has configured i3blocks to look at a specific disk, presumably he knows if
it's local or not and can make a judgement as to whether he wants this
behaviour. Consequently, we remove the -l option from df to allow remote disks
to be monitored.