mediaplayer: use playerctl when no instance is specified
If no mediaplayer instance is currently specified, mpc is used as
the default metadata utility. This results in metadata failure if
mpc/mpd are not used.

This update ensures that mpc is only used when mpd is specified in the
config file, and that when no instance is specified playerctl is used.
A side-effect of this change is that any MPRIS-compatible media player
should "just work" when playerctl is installed.

Tested locally with vlc, quodlibet and audacious.