readme: switch to AsciiDoc(tor)
AsciiDoc is a standard supporting include directives, tables and more,
conversion to many formats such as HTML and manpages and is human
readable, similarly to Markdown.

Convert the README to AsciiDoc, similarly to i3blocks.

At the same time, use the Sentence-per-line convention, which avoids
having to wrap lines and makes the document's structure and diffs
easier to read.