bandwidth2: Fix padding when specifying a label.
For reasons I don't fully understand, specifying a label (unicode in my case,
not sure if that is important) causes the leading spaces padding printed
before the current RX amount to be ignored.  This results in the entire
status line text dancing around as the number of digits changes.

For more reasons I don't fully understand, adding fallback='true' to the
pango <span> tags fixes it.