Fixed output from `lsblk` command
On my Arch machine, when getLeafDevicePaths() was calling `lsblk -spndo
NAME', it was returning lines with an unusual character that was being
included into the list of paths. I changed the final filter to the
'leaves' and added a check for that.

Also, getKernelName(path), on my machine, was getting duplicate lines
and partition paths from `lsblk -ndso KNAME` and that was messing up the
output for some reason. Removing the '-s' flag fixed it.