volume-pulseaudio: Process click inputs while in SUBSCRIBE mode.
volume-pulseaudio was previously unable to process click inputs while
being in persist mode, for valid reasons. With i3blocks now sending the
key events to stdin while in this mode (i3blocks: 6e8b51d) it is
possible to make the block interactable again.

This change moves the subscribe pipeline and printing to a background
task since stdout remains attached by default. Input events are received
on stdin which are read in a loop and passed to handle_input, that
encapsulates the case satement previously used to parse values from
$BLOCK_BUTTON.