time: Enable time/time to print its output in UTF8
This fixes i3 bar parsing errors when the output contains characters not
in the US-ASCII encoding space like the ä in the German März (March).

Perl will by default print in US-ASCII/Latin-1 encoding, where
especially the latter can contain characters that are not valid in UTF8.
If that happens, this will cause the entire parsing of the i3blocks
output to fail in i3 bar and cause it to only show an error message.