volume-pulseaudio: Restart pactl when Pulse terminates
The connection between pactl and Pulse is dropped when Pulse restarts,
and pactl is not set up to deal with this. Intead the command has to be
restarted.

This is only useful for Pulse developers and users toying with Pulse,
having to restart it without needing to restart the block/i3bar.