brightness: Stop busy waiting on Linux
This was happening on Linux after the first
call to `select` in the while loop. The
`time_to_wait` was reset to have `tv_sec == 0`
by the `select()` call and this would effectively
make us busy wait which would cause high CPU
usage.

This is how `select()` works on Linux as per
the man page: https://man7.org/linux/man-pages/man2/select.2.html